convert scores to ints
too much funny business using small floats

+++ src/lib.rs
@@ -1,10 +1,10 @@
 pub mod xor;
 pub use xor::{single_byte, xor_fixed};
 
-#[derive(Clone)]
+#[derive(Debug, Clone)]
 pub struct FreqScore {
     pub id: u8,
-    pub score: f32,
+    pub score: i32,
 }
 
 #[derive(Clone)]
